Update: 강연 컨트롤러 분리
