+++ back/src/routers/talkRouter.js
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { loginRequired } from "../middlewares/loginRequired";
-import { TalkService } from "../services/TalkService";
+import { talkReadTodayController, talkReadMyController, talkReadController, talkReadLikesRankingController } from "../contoller/talkController";
 
 const talkRouter = Router();
 
@@ -28,20 +28,7 @@ const talkRouter = Router();
  *             schema:
  *               $ref: '#/components/schemas/Talk'
  */
-talkRouter.get("/talks/today", async (req, res, next) => {
-  try {
-    const size = Number(req.query.size) < 1 ? 1 : Number(req.query.size);
-
-    const talks = await TalkService.getTodayTalk({ size });
-    if (talks.errorMessage) {
-      throw new Error(talks.errorMessage);
-    }
-
-    res.status(200).send(talks);
-  } catch (e) {
-    next(e);
-  }
-});
+talkRouter.get("/talks/today", talkReadTodayController);
 
 // 추천된 영상
 /**
@@ -67,21 +54,7 @@ talkRouter.get("/talks/today", async (req, res, next) => {
  *             schema:
  *               $ref: '#/components/schemas/Talk'
  */
-talkRouter.get("/talks/my", loginRequired, async (req, res, next) => {
-  try {
-    const size = Number(req.query.size) < 1 ? 1 : Number(req.query.size);
-    const userId = req.currentUserId;
-    const talks = await TalkService.getMyTalk({ size, userId });
-
-    if (talks.errorMessage) {
-      throw new Error(talks.errorMessage);
-    }
-
-    res.status(200).send(talks);
-  } catch (e) {
-    next(e);
-  }
-});
+talkRouter.get("/talks/my", loginRequired, talkReadMyController);
 
 // 영상 상세정보 조회
 /**
@@ -106,20 +79,7 @@ talkRouter.get("/talks/my", loginRequired, async (req, res, next) => {
  *             schema:
  *               $ref: '#components/schemas/Talk'
  */
-talkRouter.get("/talks/:talk_id", async (req, res, next) => {
-  try {
-    const id = Number(req.params.talk_id);
-
-    const talk = await TalkService.getTalk({ id });
-    if (talk.errorMessage) {
-      throw new Error(talk.errorMessage);
-    }
-
-    res.status(200).send(talk);
-  } catch (e) {
-    next(e);
-  }
-});
+talkRouter.get("/talks/:talk_id", talkReadController);
 
 /**
  * @swagger
@@ -146,15 +106,6 @@ talkRouter.get("/talks/:talk_id", async (req, res, next) => {
  *                $ref: '#components/schemas/Talk'
  */
 
-talkRouter.get("/talks/like/ranking", async function (req, res, next) {
-  try {
-    const size = Number(req.query.size);
-    const likeRanking = await TalkService.likeRanking({ size });
-
-    res.status(200).send(likeRanking);
-  } catch (error) {
-    next(error);
-  }
-});
+talkRouter.get("/talks/like/ranking", talkReadLikesRankingController);
 
 export { talkRouter };
